[REFACTOR] #189 회원 상세 컴포넌트 분리
